requirements.txt
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
cloudinary>=1.36.0
replicate>=0.25.0
